+++ animation_nodes/ui/node_menu.py
@@ -232,7 +232,6 @@ class ColorMenu(bpy.types.Menu):
         insertNode(layout, "an_SeparateColorNode", "Separate Color")
         insertNode(layout, "an_MixDataNode", "Mix", {"dataType" : repr("Color")})
         layout.separator()
-        insertNode(layout, "an_VertexColorInputNode", "Vertex Color Input")
         insertNode(layout, "an_SetVertexColorNode", "Set Vertex Color")
 
 class ListMenu(bpy.types.Menu):
@@ -373,7 +372,8 @@ class MeshMenu(bpy.types.Menu):
         layout.separator()
         insertNode(layout, "an_VertexGroupInputNode", "Vertex Group Input")
         insertNode(layout, "an_SetVertexWeightNode", "Set Vertex Weight")
-        insertNode(layout, "an_VertexColorInputNode", "Vertex Color Input")
+        insertNode(layout, "an_GetVertexColorLayerNode", "Get Vertex Color Layer")
+        insertNode(layout, "an_InsertVertexColorLayerNode", "Insert Vertex Color Layer")
         insertNode(layout, "an_SetVertexColorNode", "Set Vertex Color")
         layout.separator()
         insertNode(layout, "an_OffsetPolygonsNode", "Offset Polygons")